test(e2e): consolidate cross-chain bridge and messaging suites
Merges the token bridge trio (token_bridge_private, token_bridge_public.parallel,
token_bridge_failure_cases) into a single token_bridge.test.ts on one shared
CrossChainMessagingTest setup, converts l1_to_l2 and l2_to_l1 from .parallel
fan-out to plain files with it.each tables, and updates bootstrap.sh timeouts
and .test_patterns.yml flake patterns for the renamed files.
