+++ .test_patterns.yml
@@ -126,19 +126,19 @@ tests:
   - regex: "src/single-node/cross-chain/l1_to_l2"
     owners:
       - *palla
-  - regex: "src/single-node/cross-chain/token_bridge_private"
+  - regex: "src/single-node/cross-chain/token_bridge"
     error_regex: "✕ Claim secret is enough to consume the message"
     owners:
       - *lasse
-  - regex: "src/single-node/cross-chain/token_bridge_public"
+  - regex: "src/single-node/cross-chain/token_bridge"
     error_regex: "✕ Someone else can mint funds to me on my behalf"
     owners:
       - *lasse
-  - regex: "src/single-node/cross-chain/token_bridge_public"
+  - regex: "src/single-node/cross-chain/token_bridge"
     error_regex: "✕ Publicly deposit funds"
     owners:
       - *lasse
-  - regex: "src/single-node/cross-chain/token_bridge_failure_cases"
+  - regex: "src/single-node/cross-chain/token_bridge"
     error_regex: "✕ Can't claim funds"
     owners:
       - *lasse
